deprecate gspiceMain move to specUtils
