+++ gspice/__init__.py
@@ -1,5 +1,5 @@
-from .gspiceMain import gaussian_estimate, gp_interp
+#from .gspiceMain import gaussian_estimate, gp_interp
 from .matrixUtils import cholesky_inv, submatrix_inv, submatrix_inv_mult
-from .specUtils import standard_scale, covar, get_chimask, covar_iter_mask
+from .specUtils import standard_scale, covar, get_chimask, covar_iter_mask, gaussian_estimate, gp_interp
 from .djs_maskinterp import maskinterp1, maskinterp
 import numpy as np
